refactor(serialization): move to old dir
