+++ core/CMakeLists.txt
@@ -39,19 +39,19 @@ set(CUBOS_CORE_SOURCE
     "src/cubos/core/reflection/external/primitives.cpp"
     "src/cubos/core/reflection/external/glm.cpp"
 
-    "src/cubos/core/data/serializer.cpp"
-    "src/cubos/core/data/deserializer.cpp"
-    "src/cubos/core/data/debug_serializer.cpp"
-    "src/cubos/core/data/json_serializer.cpp"
-    "src/cubos/core/data/json_deserializer.cpp"
-    "src/cubos/core/data/binary_serializer.cpp"
-    "src/cubos/core/data/binary_deserializer.cpp"
-    "src/cubos/core/data/package.cpp"
+    "src/cubos/core/data/old/serializer.cpp"
+    "src/cubos/core/data/old/deserializer.cpp"
+    "src/cubos/core/data/old/debug_serializer.cpp"
+    "src/cubos/core/data/old/json_serializer.cpp"
+    "src/cubos/core/data/old/json_deserializer.cpp"
+    "src/cubos/core/data/old/binary_serializer.cpp"
+    "src/cubos/core/data/old/binary_deserializer.cpp"
+    "src/cubos/core/data/old/package.cpp"
     "src/cubos/core/data/fs/file.cpp"
     "src/cubos/core/data/fs/file_system.cpp"
     "src/cubos/core/data/fs/standard_archive.cpp"
     "src/cubos/core/data/fs/embedded_archive.cpp"
-    "src/cubos/core/data/context.cpp"
+    "src/cubos/core/data/old/context.cpp"
 
     "src/cubos/core/io/window.cpp"
     "src/cubos/core/io/cursor.cpp"
