+++ core/include/cubos/core/data/old/binary_deserializer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cubos/core/data/deserializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
 #include <cubos/core/memory/stream.hpp>
 
 namespace cubos::core::data
