+++ core/include/cubos/core/data/old/deserializer.hpp
@@ -4,7 +4,7 @@
 #include <unordered_map>
 #include <vector>
 
-#include <cubos/core/data/context.hpp>
+#include <cubos/core/data/old/context.hpp>
 
 namespace cubos::core::data
 {
