+++ core/include/cubos/core/data/old/package.hpp
@@ -4,8 +4,8 @@
 #include <stack>
 #include <variant>
 
-#include <cubos/core/data/deserializer.hpp>
-#include <cubos/core/data/serializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
+#include <cubos/core/data/old/serializer.hpp>
 #include <cubos/core/log.hpp>
 
 namespace cubos::core::data
