+++ core/include/cubos/core/ecs/blueprint.hpp
@@ -4,9 +4,9 @@
 
 #pragma once
 
-#include <cubos/core/data/binary_deserializer.hpp>
-#include <cubos/core/data/binary_serializer.hpp>
-#include <cubos/core/data/serialization_map.hpp>
+#include <cubos/core/data/old/binary_deserializer.hpp>
+#include <cubos/core/data/old/binary_serializer.hpp>
+#include <cubos/core/data/old/serialization_map.hpp>
 #include <cubos/core/ecs/commands.hpp>
 #include <cubos/core/memory/buffer_stream.hpp>
 #include <cubos/core/memory/type_map.hpp>
