+++ core/include/cubos/core/geom/capsule.hpp
@@ -4,8 +4,8 @@
 
 #pragma once
 
-#include <cubos/core/data/deserializer.hpp>
-#include <cubos/core/data/serializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
+#include <cubos/core/data/old/serializer.hpp>
 
 namespace cubos::core::geom
 {
