+++ core/samples/logging/main.cpp
@@ -5,7 +5,7 @@
 /// [Logging include]
 
 /// [Debug wrapper include]
-#include <cubos/core/data/debug_serializer.hpp>
+#include <cubos/core/data/old/debug_serializer.hpp>
 
 using namespace cubos::core::data;
 /// [Debug wrapper include]
