+++ core/tests/data/context.cpp
@@ -1,6 +1,6 @@
 #include <doctest/doctest.h>
 
-#include <cubos/core/data/context.hpp>
+#include <cubos/core/data/old/context.hpp>
 
 #include "utils.hpp"
