+++ engine/include/cubos/engine/collisions/aabb.hpp
@@ -6,8 +6,8 @@
 
 #include <glm/glm.hpp>
 
-#include <cubos/core/data/deserializer.hpp>
-#include <cubos/core/data/serializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
+#include <cubos/core/data/old/serializer.hpp>
 #include <cubos/core/geom/box.hpp>
 #include <cubos/core/log.hpp>
