+++ engine/src/cubos/engine/assets/asset.cpp
@@ -1,5 +1,5 @@
-#include <cubos/core/data/deserializer.hpp>
-#include <cubos/core/data/serializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
+#include <cubos/core/data/old/serializer.hpp>
 #include <cubos/core/log.hpp>
 
 #include <cubos/engine/assets/asset.hpp>
