+++ engine/src/cubos/engine/input/bindings.cpp
@@ -1,5 +1,5 @@
-#include <cubos/core/data/deserializer.hpp>
-#include <cubos/core/data/serializer.hpp>
+#include <cubos/core/data/old/deserializer.hpp>
+#include <cubos/core/data/old/serializer.hpp>
 
 #include <cubos/engine/input/bindings.hpp>
