+++ engine/src/cubos/engine/settings/plugin.cpp
@@ -1,6 +1,6 @@
 #include <cubos/core/data/fs/file_system.hpp>
 #include <cubos/core/data/fs/standard_archive.hpp>
-#include <cubos/core/data/json_deserializer.hpp>
+#include <cubos/core/data/old/json_deserializer.hpp>
 
 #include <cubos/engine/settings/plugin.hpp>
