+++ tools/quadrados/src/convert.cpp
@@ -3,8 +3,8 @@
 #include <iostream>
 #include <unordered_map>
 
-#include <cubos/core/data/binary_deserializer.hpp>
-#include <cubos/core/data/binary_serializer.hpp>
+#include <cubos/core/data/old/binary_deserializer.hpp>
+#include <cubos/core/data/old/binary_serializer.hpp>
 #include <cubos/core/log.hpp>
 #include <cubos/core/memory/endianness.hpp>
 #include <cubos/core/memory/standard_stream.hpp>
